fix(normalize): stop a repeat's tract swallowing a sibling's junction
`demote_repeats_spanning_siblings` re-spells a repeat back to the plain edit it
grew from when the tract-wide span it canonicalised to would cover a sibling.
It asked only whether a sibling *claims bases*:

    reference  ("ACGT" x 64) + "ATACAGAAAATCAGGGCATA" + ("ACGT" x 64)
    g.[261_262insGA;263_264insAA]  ->  g.[262_263dup;263_266A[6]]

The duplication claims no bases -- it occupies the junction at 263 -- so the
demotion did not fire, and the repeat kept a span (262-266) containing that
junction. The members overlap and the pair has no well-defined resulting
sequence: `parse_hgvs` accepts the output, the SPDI apply oracle declines it.

A copy count over `262_266` cannot express another member adding sequence in
the middle of those bases, so a junction strictly inside the tract is swallowed
exactly as claimed bases are. Count it.

The 3' end stays exclusive. A junction at the tract's end is *flush against* it
rather than inside it, which is the adjacency the collapse pass exists to catch
(#999, #1135) -- the same strictness `detect_insertion_overlaps` uses for an
interior junction (#1276). Demoted, the repeat becomes a duplication that clears
the sibling: `g.[262_263dup;265_266dup]`.

The indel confluence property found this and #1286; with both closed it
immediately surfaced a third defect, filed as #1290 -- a junction shifting past
*another junction*, reordering two insertions and changing the sequence.
Verified pre-existing (identical output on this branch's base), so it was masked
behind the two overlap defects rather than introduced here. The property stays

`the_ignored_indel_property_still_finds_its_defect` pinned #1287, so closing it
turns that guard red -- by design. Re-point it at #1290, the blocker the
property surfaces next and which this commit's doc comment already records.

Verified: 8930 pass, and 8930 under FERRO_ASSERT_IDEMPOTENT=1 with
FERRO_ASSERT_REPARSE=1. fmt and clippy clean.

Closes #1287
