+++ tests/it/issue_1286_shared_junction_merge.rs
@@ -29,56 +29,7 @@
 //! re-canonicalises it to a `dup` or a repeat where the reference permits.
 
 use crate::common::cis_apply_oracle::{apply, assert_normalizes_preserving, normalize};
-use crate::common::synthetic::{padded, SyntheticBuilder};
-use ferro_hgvs::spdi::hgvs_to_spdi;
-use ferro_hgvs::{parse_hgvs, HgvsVariant, Normalizer};
-
-/// Normalize against the padded synthetic contig and assert the result denotes
-/// the same bases the input did — through `hgvs_to_spdi`, a different code path
-/// from the normalizer.
-fn assert_padded_preserving(core: &str, input: &str) -> String {
-    let provider = SyntheticBuilder::genomic(core).build();
-    let normalizer = Normalizer::new(provider.clone());
-    let reference = padded(core);
-    let variant = parse_hgvs(input).expect("parse");
-    let output = normalizer
-        .normalize(&variant)
-        .expect("normalize")
-        .to_string();
-
-    let denote = |descriptor: &str| -> Option<String> {
-        let members: Vec<HgvsVariant> = match parse_hgvs(descriptor).ok()? {
-            HgvsVariant::Allele(allele) => allele.variants.clone(),
-            single => vec![single],
-        };
-        let mut triples = Vec::new();
-        for member in &members {
-            triples.push(hgvs_to_spdi(member, &provider).ok()?);
-        }
-        triples.sort_by_key(|t| std::cmp::Reverse(t.position));
-        let mut edited = reference.as_bytes().to_vec();
-        let mut claimed = reference.len();
-        for triple in &triples {
-            let start = usize::try_from(triple.position).ok()?;
-            let end = start.checked_add(triple.deletion.len())?;
-            if end > claimed {
-                return None; // overlapping members
-            }
-            edited.splice(start..end, triple.insertion.bytes());
-            claimed = start;
-        }
-        String::from_utf8(edited).ok()
-    };
-
-    let from_input = denote(input).expect("input applies");
-    let from_output = denote(&output)
-        .unwrap_or_else(|| panic!("`{input}` -> `{output}` has no resulting sequence"));
-    assert_eq!(
-        from_output, from_input,
-        "`{input}` -> `{output}` changed the sequence"
-    );
-    output
-}
+use crate::common::synthetic::assert_padded_preserving;
 
 #[test]
 fn two_insertions_reaching_one_junction_merge() {
